Add python3.4 and homeassistant task

+++ fabrecipes/debian/homeassistant/__init__.py
@@ -15,40 +15,53 @@ from fabtools.files import watch, is_dir, is_link, is_file
    Install home assistant in debian wheezy and Raspberry PI wheezy
 """
 
-
 @task
 def install():
     """
     Install home assistant in debian wheezy and Raspberry PI wheezy
     """
 
-    # Requirement platform
-    require.deb.packages([
-        'libncurses-dev', 'libreadline-dev', 'tk-dev',
-        'libsqlite3-dev', 'sqlite3',
-        'libgdbm-dev', 'libssl-dev',
-        'libbz2-dev', 'libexpat1-dev', 'liblzma-dev', 'zlib1g-dev',
-        'wget',
-    ])
+    python34()
+    homeassistant()
 
-    # Download python3
-    getfile = 'Python-3.4.3.tar.xz'
-    dst = '/tmp/%s' % getfile
-    if not is_file(dst):
-        run_as_root(
-            'wget "https://www.python.org/ftp/python/3.4.3/%s" -O %s' % (
-                getfile,
-                dst,
+@task
+def homeassistant():
+    run_as_root('pip3.4 install homeassistant')
+
+@task
+def python34():
+    """
+    Compile and install python34
+    """
+
+    if not is_file('/usr/local/bin/python3.4'):
+        # Requirement platform
+        require.deb.packages([
+            'libncurses-dev', 'libreadline-dev', 'tk-dev',
+            'libsqlite3-dev', 'sqlite3',
+            'libgdbm-dev', 'libssl-dev',
+            'libbz2-dev', 'libexpat1-dev', 'liblzma-dev', 'zlib1g-dev',
+            'wget',
+        ])
+
+        # Download python3
+        getfile = 'Python-3.4.3.tar.xz'
+        dst = '/tmp/%s' % getfile
+        if not is_file(dst):
+            run_as_root(
+                'wget "https://www.python.org/ftp/python/3.4.3/%s" -O %s' % (
+                    getfile,
+                    dst,
+                )
             )
-        )
-    with cd('/tmp'):
-        run_as_root('tar xvfJ Python-3.4.3.tar.xz')
+        with cd('/tmp'):
+            run_as_root('tar xvfJ Python-3.4.3.tar.xz')
 
-    # Install python
-    with cd('/tmp/Python-3.4.3'):
-        run_as_root('./configure')
-        run_as_root('make')
-        run_as_root('make install')
+        # Install python
+        with cd('/tmp/Python-3.4.3'):
+            run_as_root('./configure')
+            run_as_root('make')
+            run_as_root('make install')
 
-    run_as_root('rm -rf "/tmp/Python-3.4.3"')
-    run_as_root('rm "/tmp/Python-3.4.3.tar.xz"')
+        run_as_root('rm -rf "/tmp/Python-3.4.3"')
+        run_as_root('rm "/tmp/Python-3.4.3.tar.xz"')
